[CIT-1833][Lime][Android] Fix detecting touch positions on samsung devices updated to UI 3.0. See:
https://forum.developer.samsung.com/t/android-11-motionevent-differences-in-touch-positions-when-game-plugins-app-is-set-to-max-fps/11512

When Game Launcher on some Samsung devices (Android 11 / One UI 3.0+) decides that an application (by package id) is a game and the profile 
is set to "Game Booster" (Max FPS, WQHD+) there's a bug that is being fixed by this commit. I have no idea why, but it works.

To reproduce you can use any available package id which is known to be identified as a game by Game Launcher.